Import bug fix. Increment version.

+++ symple-main/src/main/java/org/symfound/controls/system/SettingsManagerControl.java
@@ -14,9 +14,10 @@ import org.symfound.controls.ConfirmableControl;
  * @author Javed Gangjee
  */
 public abstract class SettingsManagerControl extends ConfirmableControl {
-
+    
     private static final String NAME = SettingsManagerControl.class.getName();
     private static final Logger LOGGER = Logger.getLogger(NAME);
+
     /**
      *
      */
@@ -24,15 +25,18 @@ public abstract class SettingsManagerControl extends ConfirmableControl {
         super("settings-button");
     }
     
-    
     public void deleteMasterFile() {
         String masterFile = getUser().getContent().getHomeFolder() + "/Documents/SYMPLE/Settings/Master.xml";
         File file = new File(masterFile);
-        if (file.delete()) {
-            LOGGER.info("Master file " + masterFile + " deleted successfully");
+        if (file.exists()) {
+            if (file.delete()) {
+                LOGGER.info("Master file " + masterFile + " deleted successfully");
+            } else {
+                LOGGER.fatal("Failed to delete master file " + masterFile);
+            }
         } else {
-            LOGGER.fatal("Failed to delete master file " + masterFile);
+            LOGGER.warn("Master file does not exist or has already been deleted");
         }
     }
-
+    
 }
